refactor: 디렉터리명 통일
