+++ src/main/java/ex/sample/domain/user/domain/UserRole.java
@@ -1,4 +1,4 @@
-package ex.sample.domain.user.model;
+package ex.sample.domain.user.domain;
 
 
 import lombok.Getter;
